git/requirements.txt
pip install huggingface-hub
aiogram==3.10.0
asyncpg==0.29.0
openai==1.12.0
python-dotenv==1.0.0
apscheduler==3.10.4
pytz==2024.1